Fix tests for L2T, GV phase, from adding validation set
