+++ dwi_ml/training/projects/transformer_trainer.py
@@ -9,7 +9,7 @@ import torch
 from dwi_ml.tracking.io_utils import prepare_tracking_mask
 from dwi_ml.tracking.propagation import propagate_multiple_lines
 
-from dwi_ml.training.with_generation.trainer import \
+from dwi_ml.training.trainers_withGV import \
     DWIMLTrainerForTrackingOneInput
